Make changes required by the reviewer
